fix: Deprecate flat `list.gather` and mark as elementwise
Related to #22149.

This makes the only valid signature for `list.gather`: `(List(T), List(Int)) ->
List(T)` and deprecates the non-elementwise `(List(T), Int) -> List(T)` by
adding an implicit implode for now. **NOTE**: This does not affect
`list.gather([0, 1])` because it now (as is the case in the rest of polars)
gets regarded as a list scalar.