chore: Add `proptest` testing for for parquet decoding kernels
This was a bit of an experiment and learning project for me to see how
`proptest` works. I want to see if it would be relatively simple to apply to
testing of internals. I think the experiment was quite a success.

This PR just tests adds tests for some of the harder kernels from the parquet
decoding.

Some of the takeaways:
- It is surprising to see how much faster it is then python hypothesis testing.
- We should quite easily be able define strategies for all our arrow array types.
- We circumvent the problems of `cargo-fuzz` where we start having to expose
  all of our internals and we don't need additional CI steps.

Overall, I would say this was a great success. We can expand on this.