fix: pass DeltaTable._storage_options to _get_delta_lake_table if no storage_options are explicetely provided
While upgrading to `deltalake=1.0.2` and `dagster-delta==0.5.0`,
realized that `pl.scan_delta(DeltaTable).sql().collect()` would not
inherit the storage_options from the DeltaTable definition but would
instead require storage_options to be passes explicitely.

To solve this I have passed the storage options from the table to the
local scope in case of the source being a DeltaTable and the
storage_options not being explicitely provided to the `pl.scan_delta()`
function.