fix: Correct planner output schema for `join_asof`
Fixes #24000.

The problem was actually larger than the original issue showed. The columns
should also be out of order in the schema in certain situations.

This PR simplifies the code, resolves the order issues, ensures that the schema
datatype is the datatype of the original schema not the datatype of the `on`
expression, and removes a quadratic loop.