Fix edge case that happened when lazy imports interacted with NumPy.
When lazy imports were previously used, missing attributes would still call the
lazy __getattr__, which would try to set a value on self._globals. Most modules
didn't have that, so you got an AttributeError as expected, accidentally. numpy
does have a submodule called _globals, so it was trying to add an entry to
numpy._globals as if it were a dictionary, which blew up.