perf: Stop caching Parquet metadata after 8 files
This PR adds the `POLARS_MAX_CACHED_METADATA_SCANS` environment variable to
control the amount of cached Parquet metadata structs in the IR plan.

This can massively reduce the memory consumption for queries with a lot of
scans, but will harm performance as the metadata will need to be resolved
twice.

With 100 large file scans this was the outcome:

```shell
$ export POLARS_MAX_CACHED_METADATA_SCANS=1
$ time -f %M python3 100-scans.py
278436  # =  271 MB
$ export POLARS_MAX_CACHED_METADATA_SCANS=0 # turn off pruning
$ time -f %M python3 100-scans.py
3331172 # = 3253 MB
$ export POLARS_MAX_CACHED_METADATA_SCANS=8 # default value
$ time -f %M python3 100-scans.py
503148  # =  491 MB
```

This is a temporary solution until we can probably compress the metadata
somehow.

Fixes #22366.