perf: Lower `unique` to native streaming and speed up `n_unique`
This PR creates `AmortizedUnique` kernels that speed up and reduce the memory
consumption of `unique` and `n_unique` in a `group_by` context.

For slice groups, like the new streaming sorted group_by, `rolling` and
`{arr,list}.{eval,agg}`, this can speed up `unique` and `n_unique` by 10-20x.

For the in-memory group-by the difference is much more dependent on the
amount of keys and the locality of indices. In my benchmark, if can either be a
lot faster or a few percent worse on certain pathological cases. In all cases,
it drastically reduces memory consumption.